backend/requirements.txt
fastapi
uvicorn
sqlalchemy
jose
requests
python-dotenv
asyncpg
sqlalchemy[asyncio]
databases
permit
python-jose
passlib
pydantic
pyjwt